first crack at GTS regional metrics website
