+++ gts/GTS_regional_totals_FY2018_Q2.csv
@@ -1,0 +1,4 @@
+date,met,wave,total
+2018-01,701780,63624,765404
+2018-02,676688,61886,738574
+2018-03,759916,69014,828930
