+++ gts/GTS_regional_totals_FY2018_Q3.csv
@@ -1,0 +1,4 @@
+date,met,wave,total
+2018-04,773482,75758,849240
+2018-05,890444,155768,1046212
+2018-06,936078,199180,1135258
