+++ gts/GTS_regional_totals_FY2018_Q4.csv
@@ -1,0 +1,4 @@
+date,met,wave,total
+2018-07,968256,219828,1188084
+2018-08,1005594,219026,1224620
+2018-09,982458,209388,1191846
