+++ gts/GTS_regional_totals_FY2019_Q1.csv
@@ -1,0 +1,4 @@
+date,met,wave,total
+2018-10,943064,163854,1106918
+2018-11,839958,85022,924980
+2018-12,772470,64300,836770
