+++ gts/GTS_regional_totals_FY2019_Q2.csv
@@ -1,0 +1,4 @@
+date,met,wave,total
+2019-01,984466,78358,1062824
+2019-02,895986,67902,963888
+2019-03,892004,70582,962586
