+++ gts/GTS_regional_totals_FY2019_Q3.csv
@@ -1,0 +1,4 @@
+date,met,wave,total
+2019-04,843884,81494,925378
+2019-05,981844,169718,1151562
+2019-06,947972,199594,1147566
