+++ gts/GTS_regional_totals_FY2019_Q4.csv
@@ -1,0 +1,4 @@
+date,met,wave,total
+2019-07,1041212,220082,1261294
+2019-08,1043024,222676,1265700
+2019-09,992998,204256,1197254
