+++ gts/GTS_regional_totals_FY2020_Q1.csv
@@ -1,0 +1,4 @@
+date,met,wave,total
+2019-10,1006928,185542,1192470
+2019-11,877344,104654,981998
+2019-12,836598,77896,914494
