+++ gts/GTS_regional_totals_FY2020_Q2.csv
@@ -1,0 +1,4 @@
+date,met,wave,total
+2020-01,803044,80102,883146
+2020-02,770202,74524,844726
+2020-03,869432,78372,947804
