+++ gts/GTS_regional_totals_FY2020_Q3.csv
@@ -1,0 +1,4 @@
+date,met,wave,total
+2020-04,877658,95432,973090
+2020-05,887088,106754,993842
+2020-06,910610,150018,1060628
