+++ gts/GTS_regional_totals_FY2020_Q4.csv
@@ -1,0 +1,4 @@
+date,met,wave,total
+2020-07,895840,163476,1059316
+2020-08,965234,175694,1140928
+2020-09,955778,169062,1124840
