+++ gts/GTS_regional_totals_FY2021_Q1.csv
@@ -1,0 +1,4 @@
+date,met,wave,total
+2020-10,950032,160174,1110206
+2020-11,804544,83802,888346
+2020-12,812074,73062,885136
