+++ gts/GTS_regional_totals_FY2021_Q2.csv
@@ -1,0 +1,4 @@
+date,met,wave,total
+2021-01,819092,70784,889876
+2021-02,712442,61024,773466
+2021-03,530520,44916,575436
