+++ gts/GTS_regional_totals_FY2021_Q3.csv
@@ -1,0 +1,4 @@
+date,met,wave,total
+2021-04,830524,85274,915798
+2021-05,872416,146118,1018534
+2021-06,892388,193460,1085848
