+++ gts/GTS_regional_totals_FY2021_Q4.csv
@@ -1,0 +1,4 @@
+date,met,wave,total
+2021-07,1000690,227112,1227802
+2021-08,967746,226282,1194028
+2021-09,923172,205020,1128192
