+++ gts/GTS_regional_totals_FY2022_Q1.csv
@@ -1,0 +1,4 @@
+date,met,wave,total
+2021-10,925680,180742,1106422
+2021-11,825422,88858,914280
+2021-12,889904,81266,971170
