+++ gts/GTS_regional_totals_FY2022_Q2.csv
@@ -1,0 +1,4 @@
+date,met,wave,total
+2022-01,805658,62328,867986
+2022-02,739068,61764,800832
+2022-03,825214,72142,897356
