+++ gts/GTS_regional_totals_FY2022_Q3.csv
@@ -1,0 +1,4 @@
+date,met,wave,total
+2022-04,785696,78488,864184
+2022-05,876870,127808,1004678
+2022-06,923462,181566,1105028
