+++ gts/GTS_regional_totals_FY2022_Q4.csv
@@ -1,0 +1,4 @@
+date,met,wave,total
+2022-07,956888,200628,1157516
+2022-08,973248,209572,1182820
+2022-09,1145644,213980,1359624
