+++ gts/GTS_regional_totals_FY2023_Q1.csv
@@ -1,0 +1,4 @@
+date,met,wave,total
+2022-10,1169260,197302,1366562
+2022-11,1062068,115776,1177844
+2022-12,1032350,95188,1127538
