+++ gts/GTS_regional_totals_FY2023_Q2.csv
@@ -1,0 +1,4 @@
+date,met,wave,total
+2023-01,1002832,90266,1093098
+2023-02,885220,77592,962812
+2023-03,997896,82118,1080014
