+++ website/deploy/index.html
@@ -25,7 +25,7 @@
         </div>
         <div class = "divider"></div>
      <div class="mywrap">
-        <button class = "main_button" onclick="location.href='./browse.html';" value="Browse Tools" >Browse Tools</button>
+        <button class = "main_button" onclick="location.href='./gts_regional.html';" value="Browse GTS Regional Metrics" >Browse GTS Regional Metrics</button>
         &nbsp;&nbsp;&nbsp;
         <button class = "main_button" onclick="location.href='/resources.html';" value="Explore Resources" >Explore Resources</button>
         &nbsp;&nbsp;&nbsp;
