+++ website/deploy/static/main.css
@@ -367,8 +367,8 @@ button.icon_small[disabled] {
 
 .box2,
 .box {
-  width: 200px;
-  height: 400px;
+  width: 325px;
+  height: 150px;
   border: 1px solid black;
   /*border-width: 0 1px 0 1px;*/
   margin-left: 40px;
